requirements.txt
Flask
google-cloud-vision
google-cloud-datastore
pillow